- Change color setting for plot command. Can now provide a RGB color using Graph::rgb functions - Color and label available for projections and soc decomposition in DOS plot function

+++ include/plot/gnuplot.hpp
@@ -79,7 +79,7 @@ class Gnuplot : public Graph {
      * @param labels The labels corresponding to the y quantities.
      * @param colors a vector containing the color number for each vector y
      */
-    virtual void plot(const std::vector<double> &x, const std::list<std::vector<double>> &y, const std::list<std::string> &labels, const std::vector<short> &colors);
+    virtual void plot(const std::vector<double> &x, const std::list<std::vector<double>> &y, const std::list<std::string> &labels, const std::vector<unsigned> &colors);
     
     /** 
      * Plot several quantities on the screen
@@ -95,7 +95,7 @@ class Gnuplot : public Graph {
      * @param xy A list of (x,y) pairs to plot
      * @param labels The labels corresponding to the y quantities.
      */
-    virtual void plot(const std::list<std::pair<std::vector<double>,std::vector<double>>> &xy, const std::list<std::string> &labels, const std::vector<short> &colors);
+    virtual void plot(const std::list<std::pair<std::vector<double>,std::vector<double>>> &xy, const std::list<std::string> &labels, const std::vector<unsigned> &colors);
 
     /**
      * Save the graph
